Update Java CI Workflow and fix DatabasePerTenantSpec.groovy
* Update Java CI workflow; add env RUN_TESTS_ONLY which would prevent publishing when set to TRUE.
* Move test classes to individual files.
