+++ .github/workflows/gradle.yml
@@ -4,6 +4,7 @@ on:
     branches:
       - master
       - '[7-9]+.[1-9]+.x'
+      - fixTests
   pull_request:
     branches:
       - master
@@ -17,6 +18,7 @@ jobs:
     env:
       WORKSPACE: ${{ github.workspace }}
       GRADLE_OPTS: -Xmx1500m -Dfile.encoding=UTF-8
+      RUN_TESTS_ONLY: true
     steps:
       - uses: actions/checkout@v2
       - uses: actions/cache@v2
@@ -33,17 +35,18 @@ jobs:
         run: |
           [ -f ./setup.sh ] && ./setup.sh || true
       - name: Run Tests
-        run: ./gradlew check
+        run: |
+          ./gradlew --no-daemon --refresh-dependencies clean check
       - name: Publish Test Report
         uses: scacap/action-surefire-report@v1
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
           report_paths: '**/build/test-results/test/TEST-*.xml'
       - name: Run Assemble
-        if: success() && github.event_name == 'push' && matrix.java == '8'
+        if: success() && github.event_name == 'push' && matrix.java == '8' && env.RUN_TESTS_ONLY == false
         run: ./gradlew assemble
       - name: Publish to repo.grails.org
-        if: success() && github.event_name == 'push' && matrix.java == '8'
+        if: success() && github.event_name == 'push' && matrix.java == '8' && env.RUN_TESTS_ONLY == false
         env:
           SECRING_FILE: ${{ secrets.SECRING_FILE }}
           SIGNING_KEY: ${{ secrets.SIGNING_KEY }}
@@ -54,15 +57,15 @@ jobs:
           echo $SECRING_FILE | base64 -d > secring.gpg
           ./gradlew -Dorg.gradle.internal.publish.checksums.insecure=true -Psigning.keyId="$SIGNING_KEY" -Psigning.password="$SIGNING_PASSPHRASE" -Psigning.secretKeyRingFile="secring.gpg" publish
       - name: Extract branch name
-        if: success() && github.event_name == 'push' && matrix.java == '8'
+        if: success() && github.event_name == 'push' && matrix.java == '8' && env.RUN_TESTS_ONLY == false
         id: extract_branch
         run: echo ::set-output name=value::${GITHUB_REF:11}
       - name: Create Snapshot Message for the Workflow Dispatch
-        if: success() && github.event_name == 'push' && matrix.java == '8'
+        if: success() && github.event_name == 'push' && matrix.java == '8' && env.RUN_TESTS_ONLY == false
         id: dispatch_message
         run: echo ::set-output name=value::{\"message\":\"New Core Snapshot $(date) - $GITHUB_SHA\"}
       - name: Invoke the Java CI workflow in GORM Hibernate5
-        if: success() && github.event_name == 'push' && matrix.java == '8'
+        if: success() && github.event_name == 'push' && matrix.java == '8' && env.RUN_TESTS_ONLY == false
         uses: benc-uk/workflow-dispatch@v1.1
         with:
           workflow: Java CI
