+++ grails-datastore-gorm-test/src/test/groovy/grails/gorm/services/multitenancy/database/Book.groovy
@@ -1,0 +1,9 @@
+package grails.gorm.services.multitenancy.database
+
+import grails.gorm.MultiTenant
+import grails.gorm.annotation.Entity
+
+@Entity
+class Book implements MultiTenant<Book> {
+    String title
+}
