+++ grails-datastore-gorm-test/src/test/groovy/grails/gorm/services/multitenancy/database/IBookService.groovy
@@ -1,0 +1,13 @@
+package grails.gorm.services.multitenancy.database
+
+import grails.gorm.multitenancy.CurrentTenant
+import grails.gorm.services.Service
+
+@CurrentTenant
+@Service(Book)
+interface IBookService {
+
+    Book saveBook(String title)
+
+    Integer countBooks()
+}
